+++ src/main/java/org/apache/hop/datavault/hopgui/file/modelgraph/HopGuiModelGraphBase.java
@@ -178,7 +178,8 @@ public abstract class HopGuiModelGraphBase extends HopGuiAbstractGraph
               if (coachableGraph != null) {
                 coachableGraph.notifyCoachModelChanged();
               }
-            });
+            },
+            coachableGraph);
 
     innerModelSash = new SashForm(outerModelSash, SWT.HORIZONTAL);
     PropsUi.setLook(innerModelSash);